move ramfsassert into manualtest lib
